+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 ## CHANGELOG
 
-### v0.2.2 2018-01-29
+### v0.2.4 2018-01-29
+
+- Added new assertions for checking the keys of nested JSON data:
+```
+Then the JSON response should have the nested key "data.user.firstName"
+# OR
+Then the JSON response should have the nested key "data:user:firstName" with the delimiter ":"
+```
+
+### v0.2.3 2018-01-29
 
 - Added new assertions for checking the values of nested JSON data:
 ```